# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Multi-Tenant Marketing Application API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /product-tags:
    get:
      operationId: listProductTags
      tags: [productTags]
      summary: List product/service tags for the current tenant
      parameters:
        - name: market_id
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of product tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductTag"
    post:
      operationId: createProductTag
      tags: [productTags]
      summary: Create a new product tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductTagBody"
      responses:
        "201":
          description: Product tag created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductTag"
  /product-tags/{id}:
    patch:
      operationId: updateProductTag
      tags: [productTags]
      summary: Update a product tag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductTagBody"
      responses:
        "200":
          description: Product tag updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductTag"
    delete:
      operationId: deleteProductTag
      tags: [productTags]
      summary: Delete a product tag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product tag deleted
  /product-tags/{id}/assets:
    get:
      operationId: listProductTagAssets
      tags: [productTags]
      summary: List asset IDs assigned to a product tag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Asset IDs
          content:
            application/json:
              schema:
                type: object
                properties:
                  assetIds:
                    type: array
                    items:
                      type: integer
                  brandAssetIds:
                    type: array
                    items:
                      type: integer
                required:
                  - assetIds
                  - brandAssetIds
    put:
      operationId: setProductTagAssets
      tags: [productTags]
      summary: Set assets assigned to a product tag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                assetIds:
                  type: array
                  items:
                    type: integer
                brandAssetIds:
                  type: array
                  items:
                    type: integer
      responses:
        "200":
          description: Assignments updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
...
  /social-accounts:
    get:
      operationId: listSocialAccounts
      tags: [socialAccounts]
      summary: List social accounts for the current tenant
      parameters:
        - name: market_id
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of social accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SocialAccount"
    post:
      operationId: createSocialAccount
      tags: [socialAccounts]
      summary: Add a social account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSocialAccountBody"
      responses:
        "201":
          description: Social account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialAccount"
  /social-accounts/{id}:
    patch:
      operationId: updateSocialAccount
      tags: [socialAccounts]
      summary: Update a social account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSocialAccountBody"
      responses:
        "200":
          description: Social account updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialAccount"
    delete:
      operationId: deleteSocialAccount
      tags: [socialAccounts]
      summary: Delete a social account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Social account deleted
...
components:
  schemas:
...
    SocialAccount:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        platform:
          type: string
        accountName:
          type: string
        socialPilotAccountId:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - platform
        - accountName
        - socialPilotAccountId
        - createdAt
    CreateSocialAccountBody:
      type: object
      properties:
        platform:
          type: string
        accountName:
          type: string
        socialPilotAccountId:
          type: string
        marketId:
          type: integer
      required:
        - platform
        - accountName
        - socialPilotAccountId
    UpdateSocialAccountBody:
      type: object
      properties:
        platform:
          type: string
        accountName:
          type: string
        socialPilotAccountId:
          type: string
...
    ProductTag:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - name
        - createdAt
    CreateProductTagBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        marketId:
          type: integer
      required:
        - name
    UpdateProductTagBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
...