# lib/api-zod/src/generated/types/createSocialAccountBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Multi-Tenant Marketing Application API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSocialAccountBody {
  platform: string;
  accountName: string;
  socialPilotAccountId: string;
  marketId?: number;
}